Code streamlining/optimization
I worked on data loading and a bit on optimizing the plotData function
to improve the performance of the visualization. All of the data,
including the lineage tree, is now loaded at once through the
initializeEmbryo() chain of callbacks. The new data structures are also
linked to reduce duplicates of data in memory. Essentially, the csvdata
objects now hold only the x, y, z, and radius data, and link to the cell
lineage tree data structure (called cellmap) for metadata about the
cell. I think this saves us a bunch of memory. In plotData() I reduced
the number of filtering loops required in the callbacks by compiling up
front the names of all the cells involved in the plotting events.
Knowing this information, the callbacks can filter tree nodes, for
example, by simply calling indexOf to search for the node name when
checking if that node should be visible. Performance is still not silky
smooth on the later time points, but it is definitely useable and way
better than it was earlier today.